split the checkout page
